+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 
+## 1.23.0
+ September 28, 2021
+- class WebPage<T>
+- class WebPageWrapper extends WebPage<WebPageWrapper>
+- Material(child: InkWell(
+
 ## 1.22.0+2
  September 27, 2021
 - _screenSize.height * (_smallScreen ? 0.01 : 0.35),